Make object validation more general
